Move migration to generator templates dir
In preparation for creating a generator for it.
