+++ spec/flipper/adapters/active_record_spec.rb
@@ -5,28 +5,7 @@ require 'flipper/spec/shared_adapter_specs'
 # Turn off migration logging for specs
 ActiveRecord::Migration.verbose = false
 
-class CreateFlipperTables < ActiveRecord::Migration
-  def self.up
-    create_table :flipper_features do |t|
-      t.string :key, null: false
-      t.timestamps null: false
-    end
-    add_index :flipper_features, :key, unique: true
-
-    create_table :flipper_gates do |t|
-      t.string :feature_key, null: false
-      t.string :key, null: false
-      t.string :value
-      t.timestamps null: false
-    end
-    add_index :flipper_gates, [:feature_key, :key, :value], unique: true
-  end
-
-  def self.down
-    drop_table :flipper_gates
-    drop_table :flipper_features
-  end
-end
+require 'generators/flipper/templates/migration'
 
 RSpec.describe Flipper::Adapters::ActiveRecord do
   subject { described_class.new }
